requirements.txt
django
djangorestframework
psycopg2-binary
drf-yasg
python-decouple
djangorestframework-simplejwt
gunicorn
whitenoise
django-cors-headers